Question #2
For some reason the code works perfectly fine sometimes and sometimes has a bug. The code works when the spacebar is pressed the first two times but when it is pressed a third time it gets stuck and rolls to the right side of the screen while flashing random colors really fast. Definitely has to do with the moveball function and I am pretty sure the problem is in the part where the code checks for collisions.